FrequencyTool: Add --no-merge option.
The merge operation can run out of memory for very large corpora.
The --no-merge options writes a .tsv.gz file per chunk. You can
process these files yourself, for example, merging them by inserting/
updating database records.

+++ tools/src/main/java/nl/inl/blacklab/tools/frequency/FrequencyTool.java
@@ -66,6 +66,7 @@ public class FrequencyTool {
         exit("Calculate term frequencies over annotation(s) and metadata field(s).\n\n" +
                 "Usage:\n\n  FrequencyTool [--gzip] INDEX_DIR CONFIG_FILE [OUTPUT_DIR]\n\n" +
                 "  --gzip       write directly to .gz file\n" +
+                "  --no-merge   don't merge chunk files, write separate tsvs instead\n" +
                 "  INDEX_DIR    index to generate frequency lists for\n" +
                 "  CONFIG_FILE  YAML file specifying what frequency lists to generate. See README.md.\n" +
                 "  OUTPUT_DIR   where to write TSV output files (defaults to current dir)\n\n");
@@ -77,13 +78,16 @@ public class FrequencyTool {
 
         // Check for options
         int numOpts = 0;
-        boolean gzip = false;
+        FreqListOutput.Type outputType = FreqListOutput.Type.TSV;
         for (String arg: args) {
             if (arg.startsWith("--")) {
                 numOpts++;
                 switch (arg) {
                 case "--gzip":
-                    gzip = true;
+                    outputType = FreqListOutput.Type.TSV_GZIP;
+                    break;
+                case "--no-merge":
+                    outputType = FreqListOutput.Type.UNMERGED_TSV_GZ;
                     break;
                 case "--help":
                     exitUsage("");
@@ -125,25 +129,25 @@ public class FrequencyTool {
             Timer t = new Timer();
 
             // Generate the frequency lists
-            makeFrequencyLists(index, config, outputDir, gzip);
+            makeFrequencyLists(index, config, outputDir, outputType);
 
             System.out.println("TOTAL TIME: " + t.elapsedDescription(true));
         }
     }
 
-    private static void makeFrequencyLists(BlackLabIndex index, Config config, File outputDir, boolean gzip) {
+    private static void makeFrequencyLists(BlackLabIndex index, Config config, File outputDir, FreqListOutput.Type outputType) {
         AnnotatedField annotatedField = index.annotatedField(config.getAnnotatedField());
         config.check(index);
         index.setCache(new SearchCacheDummy()); // don't cache results
         for (ConfigFreqList freqList: config.getFrequencyLists()) {
             Timer t = new Timer();
-            makeFrequencyList(index, annotatedField, freqList, outputDir, gzip, config);
+            makeFrequencyList(index, annotatedField, freqList, outputDir, outputType, config);
             System.out.println("  Time: " + t.elapsedDescription());
         }
     }
 
     private static void makeFrequencyList(BlackLabIndex index, AnnotatedField annotatedField, ConfigFreqList freqList,
-                                          File outputDir, boolean gzip, Config config) {
+                                          File outputDir, FreqListOutput.Type outputType, Config config) {
         String reportName = freqList.getReportName();
 
         List<String> extraInfo = new ArrayList<>();
@@ -156,12 +160,15 @@ public class FrequencyTool {
 
         if (config.isUseRegularSearch()) {
             // Skip optimizations (debug)
-            makeFrequencyListUnoptimized(index, annotatedField, freqList, outputDir, gzip, config);
+            makeFrequencyListUnoptimized(index, annotatedField, freqList, outputDir, outputType, config);
             return;
         }
 
         // Use specifically optimized CalcTokenFrequencies
         List<String> annotationNames = freqList.getAnnotations();
+        Terms[] terms = annotationNames.stream()
+                .map(name -> index.annotationForwardIndex(annotatedField.annotation(name)).terms())
+                .toArray(Terms[]::new);
         List<Annotation> annotations = annotationNames.stream().map(annotatedField::annotation).collect(Collectors.toList());
         List<String> metadataFields = freqList.getMetadataFields();
         final List<Integer> docIds = new ArrayList<>();
@@ -209,7 +216,8 @@ public class FrequencyTool {
 
                     if (isFinalRun && chunkNumber == 0) {
                         // There's only one chunk. We can skip writing intermediate file and write result directly.
-                        FreqListOutput.TSV.write(index, annotatedField, reportName, annotationNames, occurrences, outputDir, gzip);
+                        FreqListOutput.TSV.write(index, annotatedField, reportName, annotationNames, occurrences,
+                                outputDir, outputType == FreqListOutput.Type.TSV_GZIP);
                         occurrences = null;
                     } else if (groupingTooLarge || isFinalRun) {
                         // Sort our map now.
@@ -218,32 +226,37 @@ public class FrequencyTool {
                         // Write next chunk file.
                         chunkNumber++;
                         String chunkName = reportName + chunkNumber;
-                        File chunkFile = new File(tmpDir, chunkName + ".chunk");
+                        boolean tsv = outputType == FreqListOutput.Type.UNMERGED_TSV_GZ;
+                        File chunkFile = new File(tmpDir, chunkName + (tsv ? ".tsv.gz" : ".chunk"));
                         System.out.println("  Writing " + chunkFile);
-                        writeChunkFile(chunkFile, sorted, config.isCompressTempFiles());
+                        if (!tsv) {
+                            // Write chunk files, to be merged at the end
+                            writeChunkFile(chunkFile, sorted, config.isCompressTempFiles());
+                            chunkFiles.add(chunkFile);
+                        } else {
+                            // Write separate TSV file per chunk; don't merge at the end
+                            writeTsvFile(chunkFile, sorted, terms);
+                        }
                         occurrences = null; // free memory, allocate new on next iteration
-                        chunkFiles.add(chunkFile);
                     }
                 }
             }
         }
 
         // Did we write intermediate chunk files that have to be merged?
-        if (chunkNumber > 0) {
+        if (!chunkFiles.isEmpty()) {
             // Yes, merge all the chunk files. Because they are sorted, this will consume very little memory,
             // even if the final output file is huge.
-            Terms[] terms = annotationNames.stream()
-                    .map(name -> index.annotationForwardIndex(annotatedField.annotation(name)).terms())
-                    .toArray(Terms[]::new);
             MatchSensitivity[] sensitivity = new MatchSensitivity[terms.length];
             Arrays.fill(sensitivity, MatchSensitivity.INSENSITIVE);
-            mergeChunkFiles(chunkFiles, outputDir, reportName, gzip, terms, sensitivity, config.isCompressTempFiles());
-        }
+            mergeChunkFiles(chunkFiles, outputDir, reportName, outputType == FreqListOutput.Type.TSV_GZIP,
+                    terms, sensitivity, config.isCompressTempFiles());
 
-        // Remove chunk files
-        for (File chunkFile: chunkFiles) {
-            if (!chunkFile.delete())
-                System.err.println("Could not delete: " + chunkFile);
+            // Remove chunk files
+            for (File chunkFile: chunkFiles) {
+                if (!chunkFile.delete())
+                    System.err.println("Could not delete: " + chunkFile);
+            }
         }
         if (!tmpDir.delete())
             System.err.println("Could not delete: " + tmpDir);
@@ -251,9 +264,9 @@ public class FrequencyTool {
 
     private static void writeChunkFile(File chunkFile, Map<GroupIdHash, OccurrenceCounts> occurrences, boolean compress) {
         try (FileOutputStream fileOutputStream = new FileOutputStream(chunkFile)) {
-             OutputStream gzipOutputStream = compress ? new GZIPOutputStream(fileOutputStream) : fileOutputStream;
+             OutputStream outputStream = compress ? new GZIPOutputStream(fileOutputStream) : fileOutputStream;
              try {
-                 try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(gzipOutputStream)) {
+                 try (ObjectOutputStream objectOutputStream = new ObjectOutputStream(outputStream)) {
 
                      // Write keys and values in sorted order, so we can merge later
                      objectOutputStream.writeInt(occurrences.size()); // start with number of groups
@@ -268,7 +281,7 @@ public class FrequencyTool {
                  }
              } finally {
                  if (compress)
-                     gzipOutputStream.close();
+                     outputStream.close();
              }
 
         } catch (IOException e) {
@@ -276,9 +289,32 @@ public class FrequencyTool {
         }
     }
 
+    private static void writeTsvFile(File chunkFile, Map<GroupIdHash, OccurrenceCounts> occurrences, Terms[] terms) {
+        boolean compress = true;
+        try (FileOutputStream fileOutputStream = new FileOutputStream(chunkFile)) {
+            OutputStream outputStream = compress ? new GZIPOutputStream(fileOutputStream) : fileOutputStream;
+            try {
+                try (Writer w = new OutputStreamWriter(outputStream, StandardCharsets.UTF_8);
+                        CSVPrinter csv = new CSVPrinter(w, FreqListOutputTsv.TAB_SEPARATED_FORMAT)) {
+                    for (Map.Entry<GroupIdHash, OccurrenceCounts> entry: occurrences.entrySet()) {
+                        GroupIdHash key = entry.getKey();
+                        OccurrenceCounts value = entry.getValue();
+                        FreqListOutputTsv.writeGroupRecord(null, terms, csv, key, value.hits);
+                    }
+                }
+            } finally {
+                if (compress)
+                    outputStream.close();
+            }
+        } catch (IOException e) {
+            throw new RuntimeException();
+        }
+    }
+
     // Merge the sorted subgroupings that were written to disk, writing the resulting TSV as we go.
     // This takes very little memory even if the final output file is huge.
-    private static void mergeChunkFiles(List<File> chunkFiles, File outputDir, String reportName, boolean gzip, Terms[] terms, MatchSensitivity[] sensitivity, boolean chunksCompressed) {
+    private static void mergeChunkFiles(List<File> chunkFiles, File outputDir, String reportName, boolean gzip,
+            Terms[] terms, MatchSensitivity[] sensitivity, boolean chunksCompressed) {
         File outputFile = new File(outputDir, reportName + ".tsv" + (gzip ? ".gz" : ""));
         System.out.println("  Merging " + chunkFiles.size() + " chunk files to produce " + outputFile);
         try (OutputStream outputStream = new FileOutputStream(outputFile)) {
@@ -369,7 +405,8 @@ public class FrequencyTool {
     }
 
     // Non memory-optimized version
-    private static void makeFrequencyListUnoptimized(BlackLabIndex index, AnnotatedField annotatedField, ConfigFreqList freqList, File outputDir, boolean gzip, Config config) {
+    private static void makeFrequencyListUnoptimized(BlackLabIndex index, AnnotatedField annotatedField,
+            ConfigFreqList freqList, File outputDir, FreqListOutput.Type outputType, Config config) {
 
         // Create our search
         try {
@@ -379,7 +416,8 @@ public class FrequencyTool {
             for (int i = 0; i < Math.max(1, config.getRepetitions()); i++) {
                 result = search.execute();
             }
-            FreqListOutput.TSV.write(index, annotatedField, freqList, result, outputDir, gzip);
+            FreqListOutput.TSV.write(index, annotatedField, freqList, result, outputDir,
+                    outputType == FreqListOutput.Type.TSV_GZIP);
         } catch (InvalidQuery e) {
             throw new BlackLabRuntimeException("Error creating freqList " + freqList.getReportName(), e);
         }
